Add Pico loader and remove godot binary
This commit adds Pico's OpenXR loader as new plugin.

Furthermore, this commit replaces the Godot binary with a dependency to
io.github.m4gr3d:godot, to improve maintainability of this repository.

The new dependency may get updated to an official org.godot-branded
artefact if such an artefact becomes available in the future.